Uradjeno logovanje i realtime citanje sa aparata

+++ src/main/java/com/konstil/Belezenje/radnih/vremena/repository/OperacijaRepo.java
@@ -2,6 +2,10 @@ package com.konstil.Belezenje.radnih.vremena.repository;
 
 import com.konstil.Belezenje.radnih.vremena.domain.Operacija;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface OperacijaRepo extends JpaRepository<Operacija,Long> {
+import java.util.List;
+
+public interface OperacijaRepo extends JpaRepository<Operacija, Integer>{
+
 }
